feat: Integrate Header into Left Area
