+++ invokeai/frontend/web/src/app/components/App.tsx
@@ -1,14 +1,13 @@
-import { Flex, Grid } from '@chakra-ui/react';
-import { useStore } from '@nanostores/react';
+import { Flex } from '@chakra-ui/react';
+import { useSocketIO } from 'app/hooks/useSocketIO';
 import { useLogger } from 'app/logging/useLogger';
 import { appStarted } from 'app/store/middleware/listenerMiddleware/listeners/appStarted';
-import { $headerComponent } from 'app/store/nanostores/headerComponent';
 import { useAppDispatch, useAppSelector } from 'app/store/storeHooks';
 import { PartialAppConfig } from 'app/types/invokeai';
 import ImageUploader from 'common/components/ImageUploader';
+import { useClearStorage } from 'common/hooks/useClearStorage';
 import ChangeBoardModal from 'features/changeBoardModal/components/ChangeBoardModal';
 import DeleteImageModal from 'features/deleteImageModal/components/DeleteImageModal';
-import SiteHeader from 'features/system/components/SiteHeader';
 import { configChanged } from 'features/system/store/configSlice';
 import { languageSelector } from 'features/system/store/systemSelectors';
 import InvokeTabs from 'features/ui/components/InvokeTabs';
@@ -20,8 +19,6 @@ import AppErrorBoundaryFallback from './AppErrorBoundaryFallback';
 import GlobalHotkeys from './GlobalHotkeys';
 import PreselectedImage from './PreselectedImage';
 import Toaster from './Toaster';
-import { useSocketIO } from 'app/hooks/useSocketIO';
-import { useClearStorage } from 'common/hooks/useClearStorage';
 
 const DEFAULT_CONFIG = {};
 
@@ -63,25 +60,21 @@ const App = ({ config = DEFAULT_CONFIG, selectedImage }: Props) => {
     dispatch(appStarted());
   }, [dispatch]);
 
-  const headerComponent = useStore($headerComponent);
-
   return (
     <ErrorBoundary
       onReset={handleReset}
       FallbackComponent={AppErrorBoundaryFallback}
     >
-      <Grid w="100vw" h="100vh" position="relative" overflow="hidden">
+      <Flex w="100vw" h="100vh" position="relative" overflow="hidden">
         <ImageUploader>
-          <Grid
+          <Flex
             sx={{
               gap: 4,
               p: 4,
-              gridAutoRows: 'min-content auto',
               w: 'full',
               h: 'full',
             }}
           >
-            {headerComponent || <SiteHeader />}
             <Flex
               sx={{
                 gap: 4,
@@ -91,9 +84,9 @@ const App = ({ config = DEFAULT_CONFIG, selectedImage }: Props) => {
             >
               <InvokeTabs />
             </Flex>
-          </Grid>
+          </Flex>
         </ImageUploader>
-      </Grid>
+      </Flex>
       <DeleteImageModal />
       <ChangeBoardModal />
       <Toaster />
